refactor(preingestion-manager): move to its own crate
- preingestion-manager moved to its own crate
- TestMeter moved to utils under test-support feature
