+++ crates/api/src/lib.rs
@@ -89,7 +89,6 @@ mod mqtt_state_change_hook;
 mod network_segment;
 mod nvl_partition_monitor;
 mod nvlink;
-mod preingestion_manager;
 mod rack;
 mod redfish;
 mod run;
