+++ crates/api/src/setup.rs
@@ -23,6 +23,7 @@ use std::sync::Arc;
 use arc_swap::ArcSwap;
 use carbide_firmware::FirmwareDownloader;
 use carbide_ipmi::IPMITool;
+use carbide_preingestion_manager::PreingestionManager;
 use carbide_redfish::libredfish::RedfishClientPool;
 use carbide_redfish::nv_redfish::NvRedfishClientPool;
 use carbide_site_explorer::{BmcEndpointExplorer, SiteExplorer};
@@ -72,7 +73,6 @@ use crate::measured_boot::metrics_collector::MeasuredBootMetricsCollector;
 use crate::mqtt_state_change_hook::hook::MqttStateChangeHook;
 use crate::nvl_partition_monitor::NvlPartitionMonitor;
 use crate::nvlink::{NmxmClientPool, NmxmClientPoolImpl};
-use crate::preingestion_manager::PreingestionManager;
 use crate::rack::bms_client::BmsDsxExchangeHandle;
 use crate::scout_stream::ConnectionRegistry;
 use crate::state_controller::common_services::CommonStateHandlerServices;
