+++ crates/api/src/tests/service_health_metrics.rs
@@ -20,11 +20,11 @@ use std::sync::{Arc, Mutex};
 use model::resource_pool::ResourcePoolStats;
 use prometheus_text_parser::ParsedPrometheusMetrics;
 use sqlx::PgPool;
+use utils::test_support::test_meter::TestMeter;
 
 use crate::logging::service_health_metrics::{
     ServiceHealthContext, start_export_service_health_metrics,
 };
-use crate::tests::common::test_meter::TestMeter;
 
 #[crate::sqlx_test]
 async fn test_service_health_metrics(pool: PgPool) -> Result<(), Box<dyn std::error::Error>> {
