+++ crates/preingestion-manager/Cargo.toml
@@ -1,0 +1,40 @@
+[package]
+name = "carbide-preingestion-manager"
+version = "0.0.1"
+edition.workspace = true
+description = "Carbide Pre-ingestion Manager"
+license.workspace = true
+authors.workspace = true
+
+[lib]
+name = "carbide_preingestion_manager"
+
+[dependencies]
+bmc-vendor = { path = "../bmc-vendor" }
+carbide-api-db = { path = "../api-db" }
+carbide-api-model = { path = "../api-model" }
+carbide-firmware = { path = "../firmware" }
+carbide-redfish = { path = "../redfish" }
+carbide-secrets = { path = "../secrets" }
+carbide-site-explorer = { path = "../site-explorer" }
+carbide-ssh = { path = "../ssh" }
+carbide-utils = { path = "../utils" }
+
+# these are alphabetized
+chrono = { workspace = true }
+futures-util = { workspace = true }
+libredfish = { workspace = true }
+opentelemetry = { workspace = true }
+sqlx = { workspace = true }
+thiserror = { workspace = true }
+time = { workspace = true }
+tokio = { workspace = true }
+tokio-util = { workspace = true }
+tracing = { workspace = true }
+version-compare = { workspace = true }
+
+[dev-dependencies]
+prometheus-text-parser = { path = "../prometheus-text-parser" }
+
+[lints]
+workspace = true
