+++ crates/preingestion-manager/src/fixtures/test_metrics_collector.txt
