+++ crates/utils/src/test_support/mod.rs
@@ -15,14 +15,4 @@
  * limitations under the License.
  */
 
-//! Contains common functionality between integration tests
-
-pub mod api_fixtures;
-pub mod endpoint;
-pub mod ib_guid_pool;
-pub mod mac_address_pool;
-pub mod metadata;
-pub mod network_segment;
-pub mod rpc_builder;
-pub mod sqlx_fixtures;
-pub mod test_certificates;
+pub mod test_meter;
